// src/models/userModel.ts
export interface User {
    id?: number;
    nome: string;
    idade: string;
    profissao: string;
    urlImagem: string;
    urlCurriculo: string;
    habilidades: string;
    nivelSenioridade: string;
    cep: string;
    estado: string;
    cidade: string;
    rua: string;
    numero: string;
    complemento: string;
    bairro: string;
}